Add documentation and more tests
